requirements.txt
flask
transformers
torch
numpy
pandas
g4f-client
requests
bert-tokenizer-fast
